+++ __main__.py
@@ -1,3 +1,4 @@
+ 
 """
 Pukage
 Choose-your-own-adventure game.
@@ -9,3 +10,4 @@ from main import startingMenu
 
 if __name__ == "__main__":
     startingMenu()
+