+++ waittype.py
@@ -1,7 +1,7 @@
 """
 Pukage
 Choose-your-own-adventure game.
-Copyright 2020 Daniel Zhang, Jeffery Zang, Li Feng Yin
+Copyright 2020 Daniel Zhang, Jeffrey Zang, Li Feng Yin
 MIT License
 """
 
@@ -9,12 +9,12 @@ from scrolltype import scrolltype
 from time import sleep
 
 
-waitTime = 0.5
+waitTime = 0.24
 
 def waittype(string):
     scrolltype(string);
     time = len(string.split()) * waitTime
-    scrolltype("\n\n\n\n")
+    scrolltype("\n\n")
     sleep(time)